Lisa was here and adjusted combat system to enemies

+++ Assets/Tech/Scripts/Player/PlayerStats.cs
@@ -13,7 +13,7 @@ public class PlayerStats : MonoBehaviour, IHealable, IDamageable
     //Propriedades
     public float CurrentHealth => _health;
     public float MaxHealth => stats.maxHealth;
-    public int AttackDamage => stats.attackDamage;
+    public float AttackDamage => stats.attackDamage;
     public float CooldownReduction => stats.cooldownReduction;
     public float Speed => stats.moveSpeed;
